+++ src/IECoreArnold/ShaderNetworkAlgo.cpp
@@ -397,6 +397,42 @@ ShaderNetworkPtr preprocessedNetwork( const IECoreScene::ShaderNetwork *shaderNe
 	return result;
 }
 
+template<typename T>
+T *reportedCast( const IECore::RunTimeTyped *v, const char *type, const IECore::InternedString &name )
+{
+	T *t = IECore::runTimeCast<T>( v );
+	if( t )
+	{
+		return t;
+	}
+
+	IECore::msg( IECore::Msg::Warning, "IECoreArnold::ShaderNetworkAlgo", fmt::format( "Expected {} but got {} for {} \"{}\".", T::staticTypeName(), v->typeName(), type, name.c_str() ) );
+	return nullptr;
+}
+
+template<typename T>
+const T *attribute( const IECore::CompoundObject::ObjectMap &attributes, IECore::InternedString name )
+{
+	auto it = attributes.find( name );
+	if( it == attributes.end() )
+	{
+		return nullptr;
+	}
+
+	return reportedCast<const T>( it->second.get(), "attribute", name );
+}
+
+pair<IECore::InternedString, const IECoreScene::ShaderNetwork *> shaderNetworkAttribute( const vector<IECore::InternedString> &attributeNames, const IECore::CompoundObject::ObjectMap &attributes )
+{
+	for( const auto &name : attributeNames )
+	{
+		if( const auto *shaderNetwork = attribute<IECoreScene::ShaderNetwork>( attributes, name ) )
+		{
+			return { name, shaderNetwork };
+		}
+	}
+	return { IECore::InternedString(), nullptr };
+}
 
 } // namespace
 
@@ -718,6 +754,26 @@ const InternedString g_widthParameter( "width" );
 const InternedString g_wrapSParameter( "wrapS" );
 const InternedString g_wrapTParameter( "wrapT" );
 
+const InternedString g_cameraVisibilityAttributeName( "ai:visibility:camera" );
+const InternedString g_diffuseReflectVisibilityAttributeName( "ai:visibility:diffuse_reflect" );
+const InternedString g_diffuseTransmitVisibilityAttributeName( "ai:visibility:diffuse_transmit" );
+const InternedString g_lightAttributeName( "light" );
+const InternedString g_shadowVisibilityAttributeName( "ai:visibility:shadow" );
+const InternedString g_specularReflectVisibilityAttributeName( "ai:visibility:specular_reflect" );
+const InternedString g_specularTransmitVisibilityAttributeName( "ai:visibility:specular_transmit" );
+const InternedString g_subsurfaceVisibilityAttributeName( "ai:visibility:subsurface" );
+const InternedString g_volumeVisibilityAttributeName( "ai:visibility:volume" );
+
+const InternedString g_emptyString( "" );
+
+const std::vector<IECore::InternedString> g_surfaceShaderAttributeNames = {
+	"ai:surface",
+	"osl:surface",
+	/// \todo Remove support for interpreting "osl:shader" as a surface shader assignment.
+	"osl:shader",
+	"surface"
+};
+
 const string g_arnoldNamespace( "arnold:" );
 
 void transferUSDLightParameters( ShaderNetwork *network, InternedString shaderHandle, const Shader *usdShader, Shader *shader )
@@ -905,6 +961,41 @@ void convertUSDUVTextures( ShaderNetwork *network )
 
 		replaceUSDShader( network, handle, std::move( imageShader ) );
 	}
+}
+
+std::pair<ShaderNetwork::Parameter, ShaderNetwork::Parameter> surfaceGlowParameters( const IECoreScene::ShaderNetwork *shaderNetwork )
+{
+	ShaderNetwork::Parameter emissionColorParameter;
+	ShaderNetwork::Parameter emissionColorInput;
+	if( !shaderNetwork )
+	{
+		return { emissionColorParameter, emissionColorInput };
+	}
+
+	for( const auto &[handle, shader] : shaderNetwork->shaders() )
+	{
+		if(
+			shader->getName() == "standard_surface" ||
+			shader->getName() == "standard_hair" ||
+			shader->getName() == "toon" ||
+			shader->getName() == "openpbr_surface"
+		)
+		{
+			emissionColorParameter = { handle, g_emissionColorParameter };
+			break;
+		}
+		else if( shader->getName() == "UsdPreviewSurface" )
+		{
+			emissionColorParameter = { handle, g_emissiveColorParameter };
+			break;
+		}
+	}
+	if( emissionColorParameter )
+	{
+		emissionColorInput = shaderNetwork->input( emissionColorParameter );
+	}
+
+	return { emissionColorParameter, emissionColorInput };
 }
 
 } // namespace
@@ -1181,6 +1272,120 @@ void hashSubstitutions( const IECoreScene::ShaderNetwork *shaderNetwork, Interne
 void applySubstitutions( IECoreScene::ShaderNetwork *shaderNetwork, InternedString attributeName, const IECore::CompoundObject *attributes )
 {
 	IECoreScene::ShaderNetworkAlgo::applyRenderAdaptors( shaderNetwork, attributeName, attributes );
+}
+
+//////////////////////////////////////////////////////////////////////////
+// USDMeshLight Conversion
+//////////////////////////////////////////////////////////////////////////
+
+ConstCompoundObjectPtr convertUSDMeshLightAttributes( const CompoundObject *attributes )
+{
+	const auto *lightNetwork = attribute<IECoreScene::ShaderNetwork>( attributes->members(), g_lightAttributeName );
+	if( !lightNetwork )
+	{
+		return attributes;
+	}
+
+	const IECoreScene::Shader *outputShader = lightNetwork->outputShader();
+	if( !outputShader || outputShader->getName() != "MeshLight" )
+	{
+		return attributes;
+	}
+
+	IECore::CompoundObjectPtr result = attributes->copy();
+
+	IECoreScene::ShaderNetworkPtr newLightShaderNetwork = lightNetwork->copy();
+	const IECoreScene::ShaderNetwork *surfaceNetwork = shaderNetworkAttribute( g_surfaceShaderAttributeNames, attributes->members() ).second;
+
+	const auto &[emissionColorParameter, emissionColorInput] = surfaceGlowParameters( surfaceNetwork );
+
+	ShaderNetwork::Parameter lightOutputParameter = lightNetwork->getOutput();
+	const Shader *lightOutputShader = lightNetwork->outputShader();
+
+	ShaderPtr newLightShader = new Shader( "mesh_light", "ai:light" );
+	transferUSDLightParameters( newLightShaderNetwork.get(), lightOutputParameter.shader, lightOutputShader, newLightShader.get() );
+
+	// The potential light inputs are in the first row of this matrix.
+	// The potential surface inputs are in the first column.
+	// The cells are the resulting mesh light color / input.
+	// C = light color x surface color. If 0 or 1 in parenthesis, it means it's known to be that value.
+	// TINT = A multiply shader combining the surface and light colors.
+	// Light / Emission Tex = the texture is connected directly without tint.
+	//                         | LightColor 0 | LightColor 0-1 | LightColor 1  | LightColor Textured
+	// EmissionColor 0         |      C(0)    |       C(0)     |     C(0)      |       C(0)
+	// EmissionColor 0-1       |      C(0)    |       C        |     C         |       TINT
+	// EmissionColor 1         |      C(0)    |       C        |     C(1)      |     Light Tex
+	// EmissionColor Textured  |      C(0)    |      TINT      |  Emission Tex |       TINT
+
+	const Color3f lightColor = parameterValue( lightOutputShader, g_colorParameter, Color3f( 1.f ) );
+	const Color3f emissionColor = emissionColorParameter ? parameterValue( surfaceNetwork->getShader( emissionColorParameter.shader ), emissionColorParameter.name, Color3f( 0.f ) ) : Color3f( 0.f );
+	if( emissionColorParameter )
+	{
+		newLightShader->parameters()[g_colorParameter] = new Color3fData( emissionColor * lightColor );
+	}
+
+	InternedString tintHandle;
+	const ShaderNetwork::Parameter meshLightColorParameter = { lightOutputParameter.shader, g_colorParameter };
+	ShaderNetwork::Parameter meshLightColorInput = newLightShaderNetwork->input( meshLightColorParameter );
+	// Remove the input to the light color. We will add it back later if needed.
+	removeInput( newLightShaderNetwork.get(), meshLightColorParameter );
+
+	if( emissionColorInput && ( lightColor != Color3f( 0.f ) || meshLightColorInput ) )
+	{
+		ShaderNetworkPtr glowNetwork = surfaceNetwork->copy();
+		glowNetwork->setOutput( emissionColorInput );
+		IECoreScene::ShaderNetworkAlgo::removeUnusedShaders( glowNetwork.get() );
+		ShaderNetwork::Parameter newGlowColorInput = IECoreScene::ShaderNetworkAlgo::addShaders( newLightShaderNetwork.get(), glowNetwork.get(), /* connections = */ true );
+
+		if( lightColor != Color3f( 1.f ) || meshLightColorInput )
+		{
+			ShaderPtr tintShader = new Shader( "multiply", "ai:surface", { { "input2", new Color3fData( lightColor ) } } );
+			tintHandle = newLightShaderNetwork->addShader( InternedString( "tint" ), std::move( tintShader ) );
+
+			newLightShaderNetwork->addConnection( { newGlowColorInput, { tintHandle, "input1" } } );
+			newLightShaderNetwork->addConnection( { { tintHandle, "out" }, meshLightColorParameter } );
+		}
+		else
+		{
+			newLightShaderNetwork->addConnection( { newGlowColorInput, meshLightColorParameter } );
+		}
+	}
+
+	if( meshLightColorInput && ( emissionColor != Color3f( 0.f ) || emissionColorInput ) )
+	{
+		if( emissionColor != Color3f( 1.f ) || emissionColorInput )
+		{
+			if( tintHandle == g_emptyString )
+			{
+				ShaderPtr tintShader = new Shader( "multiply", "ai:surface", { { "input1", new Color3fData( emissionColor ) } } );
+				tintHandle = newLightShaderNetwork->addShader( InternedString( "tint" ), std::move( tintShader ) );
+
+				newLightShaderNetwork->addConnection( { { tintHandle, "out" }, meshLightColorParameter } );
+			}
+
+			newLightShaderNetwork->addConnection( { meshLightColorInput, { tintHandle, "input2" } } );
+		}
+		else
+		{
+			newLightShaderNetwork->addConnection( { meshLightColorInput, meshLightColorParameter } );
+		}
+	}
+
+	replaceUSDShader( newLightShaderNetwork.get(), lightOutputParameter.shader, std::move( newLightShader ) );
+	IECoreScene::ShaderNetworkAlgo::removeUnusedShaders( newLightShaderNetwork.get() );
+
+	result->members()[g_lightAttributeName] = std::move( newLightShaderNetwork );
+
+	result->members().try_emplace( g_cameraVisibilityAttributeName, new IECore::BoolData( true ) );
+	result->members().try_emplace( g_shadowVisibilityAttributeName, new IECore::BoolData( false ) );
+	result->members().try_emplace( g_diffuseReflectVisibilityAttributeName, new IECore::BoolData( false ) );
+	result->members().try_emplace( g_specularReflectVisibilityAttributeName, new IECore::BoolData( false ) );
+	result->members().try_emplace( g_diffuseTransmitVisibilityAttributeName, new IECore::BoolData( false ) );
+	result->members().try_emplace( g_specularTransmitVisibilityAttributeName, new IECore::BoolData( false ) );
+	result->members().try_emplace( g_volumeVisibilityAttributeName, new IECore::BoolData( false ) );
+	result->members().try_emplace( g_subsurfaceVisibilityAttributeName, new IECore::BoolData( false ) );
+
+	return result;
 }
 
 } // namespace IECoreArnold::ShaderNetworkAlgo